+++ FA2sp/Miscs/SaveMap.h
@@ -4,6 +4,9 @@
 
 #include "../FA2sp.h"
 
+#include <optional>
+#include <filesystem>
+
 class SaveMapExt
 {
 private:
@@ -11,6 +14,7 @@ private:
 public:
     static bool IsAutoSaving;
     static ppmfc::CString FileName;
+    static std::optional<std::filesystem::file_time_type> SaveTime;
 
     static void ResetTimer();
     static void StopTimer();