+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Multiselection deselect all hotkey changed from CTRL+D into CTRL+SHIFT+D
 - Added hotkey CTRL+SHIFT+F for Navigate to coordinate
 - SliderCtrl now displays tooltip indicating current strength of the object
+- Now Map Editor will hint if the map file had been externally modified. Can be translated by `FileWatcherMessage`.
 
 ## RELEASE 1.6.1 (2023-05-06)
 - New Map Tool: Navigate to coordinate, can be translated by `Menu.MapTools.NavigateCoordinate`, more locolization tags could be found in the document